(or at least `c++2a`) and fix some clazy warning

Get MSVC to properly set the `__cplusplus` value (sigh...)

Fix MSVC compiler and clazy warnings
reg.: operator '|': deprecated between enumerations of different types (C5054)
some of which only in 32-bit or Qt 5.9 builds.

Fix GCC compiler and clazy warnings
reg.: implicit capture of ‘this’ via ‘[=]’ is deprecated in C++20 [-Wdeprecated]
reg.: 'Qt::MidButton' is deprecated: MidButton is deprecated. Use MiddleButton instead [-Wdeprecated-declarations]
reg.: compound assignment with ‘volatile’-qualified left operand is deprecated [-Wvolatile]
reg.: 'QTextStream& QTextStreamFunctions::endl(QTextStream&)' is deprecated: Use Qt::endl [-Wdeprecated-declarations]

Silence/disable the remaining MSVC compiler warnings
reg.: operator '&': deprecated between enumerations of different types (C5054) in Qt and and reg.:'register' is no longer a supported storage class (C5022) in Jack, by disabling them.

Using a C++17 feature
idea stolen from exportxml.cpp

+++ awl/poslabel.cpp
@@ -30,7 +30,7 @@ PosLabel::PosLabel(TempoMap* tl, TimeSigMap* sl, QWidget* parent)
    : QLabel(parent), pos(tl, sl)
       {
       _smpte = false;
-      setFrameStyle(WinPanel | Sunken);
+      setFrameStyle(static_cast<int>(WinPanel) | static_cast<int>(Sunken));
       setLineWidth(2);
       setMidLineWidth(3);
       int fw = style()->pixelMetric(QStyle::PM_DefaultFrameWidth);
@@ -42,7 +42,7 @@ PosLabel::PosLabel(QWidget* parent)
    : QLabel(parent)
       {
       _smpte = false;
-      setFrameStyle(WinPanel | Sunken);
+      setFrameStyle(static_cast<int>(WinPanel) | static_cast<int>(Sunken));
       setLineWidth(2);
       setMidLineWidth(3);
       int fw = style()->pixelMetric(QStyle::PM_DefaultFrameWidth);
